+++ legend-engine-xts-persistence/legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-logical-plan/src/main/java/org/finos/legend/engine/persistence/components/ingestmode/versioning/DeriveDataErrorRowsLogicalPlan.java
@@ -1,4 +1,4 @@
-// Copyright 2023 Goldman Sachs
+// Copyright 2024 Goldman Sachs
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,29 +14,28 @@
 
 package org.finos.legend.engine.persistence.components.ingestmode.versioning;
 
-import org.finos.legend.engine.persistence.components.common.DedupAndVersionErrorSqlType;
 import org.finos.legend.engine.persistence.components.logicalplan.LogicalPlan;
-import org.finos.legend.engine.persistence.components.logicalplan.datasets.*;
-import org.finos.legend.engine.persistence.components.logicalplan.values.FieldValue;
-import org.finos.legend.engine.persistence.components.logicalplan.values.FunctionImpl;
-import org.finos.legend.engine.persistence.components.logicalplan.values.FunctionName;
-import org.finos.legend.engine.persistence.components.logicalplan.values.Value;
+import org.finos.legend.engine.persistence.components.logicalplan.conditions.GreaterThan;
+import org.finos.legend.engine.persistence.components.logicalplan.datasets.Dataset;
+import org.finos.legend.engine.persistence.components.logicalplan.datasets.Selection;
+import org.finos.legend.engine.persistence.components.logicalplan.values.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class DeriveDataErrorCheckLogicalPlan implements VersioningStrategyVisitor<LogicalPlan>
+public class DeriveDataErrorRowsLogicalPlan implements VersioningStrategyVisitor<LogicalPlan>
 {
+    private List<String> primaryKeys;
+    private List<String> remainingColumns;
+    private Dataset tempStagingDataset;
+    private int sampleRowCount;
 
-    List<String> primaryKeys;
-    List<String> remainingColumns;
-    Dataset tempStagingDataset;
-
-    public DeriveDataErrorCheckLogicalPlan(List<String> primaryKeys, List<String> remainingColumns, Dataset tempStagingDataset)
+    public DeriveDataErrorRowsLogicalPlan(List<String> primaryKeys, List<String> remainingColumns, Dataset tempStagingDataset, int sampleRowCount)
     {
         this.primaryKeys = primaryKeys;
         this.remainingColumns = remainingColumns;
         this.tempStagingDataset = tempStagingDataset;
+        this.sampleRowCount = sampleRowCount;
     }
 
     @Override
@@ -50,7 +49,7 @@ public class DeriveDataErrorCheckLogicalPlan implements VersioningStrategyVisito
     {
         if (maxVersionStrategy.performStageVersioning())
         {
-            return getLogicalPlanForDataErrorCheck(maxVersionStrategy.versioningField());
+            return getLogicalPlanForDataErrors(maxVersionStrategy.versioningField());
         }
         else
         {
@@ -63,7 +62,7 @@ public class DeriveDataErrorCheckLogicalPlan implements VersioningStrategyVisito
     {
         if (allVersionsStrategyAbstract.performStageVersioning())
         {
-            return getLogicalPlanForDataErrorCheck(allVersionsStrategyAbstract.versioningField());
+            return getLogicalPlanForDataErrors(allVersionsStrategyAbstract.versioningField());
         }
         else
         {
@@ -71,10 +70,9 @@ public class DeriveDataErrorCheckLogicalPlan implements VersioningStrategyVisito
         }
     }
 
-    private LogicalPlan getLogicalPlanForDataErrorCheck(String versionField)
+    private LogicalPlan getLogicalPlanForDataErrors(String versionField)
     {
-        String maxDataErrorAlias = DedupAndVersionErrorSqlType.MAX_DATA_ERRORS.name();
-        String distinctRowCount = "legend_persistence_distinct_rows";
+        String distinctRowCount = "legend_persistence_error_count";
         List<Value> pKsAndVersion = new ArrayList<>();
         for (String pk: primaryKeys)
         {
@@ -94,22 +92,16 @@ public class DeriveDataErrorCheckLogicalPlan implements VersioningStrategyVisito
                 .alias(distinctRowCount)
                 .build();
 
-        Selection selectCountDataError = Selection.builder()
+        Selection selectDataError = Selection.builder()
                 .source(tempStagingDataset)
                 .groupByFields(pKsAndVersion)
+                .addAllFields(pKsAndVersion)
                 .addFields(countDistinct)
-                .alias(tempStagingDataset.datasetReference().alias())
+                .havingCondition(GreaterThan.of(FieldValue.builder().fieldName(distinctRowCount).build(), ObjectValue.of(1)))
+                .limit(sampleRowCount)
                 .build();
-        FunctionImpl maxCount = FunctionImpl.builder()
-                .functionName(FunctionName.MAX)
-                .addValue(FieldValue.builder().fieldName(distinctRowCount).build())
-                .alias(maxDataErrorAlias)
-                .build();
-        Selection maxDataErrorCount = Selection.builder()
-                .source(selectCountDataError)
-                .addFields(maxCount)
-                .build();
-        return LogicalPlan.builder().addOps(maxDataErrorCount).build();
+
+        return LogicalPlan.builder().addOps(selectDataError).build();
     }
 
 }
