+++ legend-engine-xts-persistence/legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-logical-plan/src/main/java/org/finos/legend/engine/persistence/components/util/TableNameGenUtils.java
@@ -1,0 +1,34 @@
+// Copyright 2024 Goldman Sachs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.finos.legend.engine.persistence.components.util;
+
+import java.util.UUID;
+
+import static org.finos.legend.engine.persistence.components.util.LogicalPlanUtils.UNDERSCORE;
+
+public class TableNameGenUtils
+{
+    private static String generateTableSuffix()
+    {
+        UUID uuid = UUID.randomUUID();
+        int uuidHashCode = Math.abs(uuid.hashCode());
+        return UNDERSCORE + "LP" + UNDERSCORE + Integer.toString(uuidHashCode, 36);
+    }
+
+    public static String generateTableName(String baseTableName, String suffix)
+    {
+        return baseTableName + UNDERSCORE + suffix + UNDERSCORE + generateTableSuffix();
+    }
+}
