+++ legend-engine-xts-persistence/legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-bigquery/src/main/java/org/finos/legend/engine/persistence/components/relational/bigquery/executor/BigQueryExecutor.java
@@ -96,8 +96,7 @@ public class BigQueryExecutor implements Executor<SqlGen, TabularData, SqlPlan>
     @Override
     public List<TabularData> executePhysicalPlanAndGetResults(SqlPlan physicalPlan, int rows)
     {
-        // TODO to be implemented
-        return null;
+        throw new RuntimeException("Not implemented for Big Query");
     }
 
     @Override
