+++ legend-engine-xts-persistence/legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-snowflake/src/main/java/org/finos/legend/engine/persistence/components/relational/snowflake/sqldom/schemaops/statements/CopyStatement.java
@@ -223,7 +223,7 @@ public class CopyStatement implements DMLStatement
     {
         if (srcTable == null)
         {
-            throw new SqlDomException("selectStatement is mandatory for Copy Table Command");
+            throw new SqlDomException("srcTable is mandatory for Copy Table Command");
         }
 
         if (table == null)
